+++ tests/test_metamod.py
@@ -11,7 +11,7 @@ from imod.couplers.metamod import MetaMod
 from imod.mf6 import open_cbc, open_hds
 from numpy.testing import assert_array_almost_equal
 from pytest_cases import parametrize_with_cases
-from test_scripts.water_balance.combine import create_modflow_waterbalance_file
+from test_scripts.mf6_water_balance.combine import create_modflow_waterbalance_file
 from test_utilities import numeric_csvfiles_equal
