--- tests/test_scripts/water_balance/readme.txt
@@ -1,4 +1,0 @@
-the python scripts in this folder can be used to generate an excel file containing the water
-balances of modflow6, metaswap and dflow.
-So it can be used for tyhe output of the dfm_metamod driver.
-The entry point is located in "combine.py"
